twitch_bot/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
attrs==24.3.0
frozenlist==1.5.0
idna==3.10
iso8601==2.1.0
multidict==6.1.0
propcache==0.2.1
python-dotenv==1.0.1
twitchio==2.10.0
typing_extensions==4.12.2
yarl==1.18.3
